Update: DetailBlogPage migration
